# Improve performance of `IsTuple()`