@jnyrup 2022-12-31
> net471/netcoreapp20/netstandard21 introduced the `ITuple` interface which all `Tuple` and `ValueTuple` types implement.

To strengthen this argument: as of writing _only_ `Tuple` and `ValueTuple` types in the BCL implement `ITuple` and this is also similar to what [Roslyn does](https://github.com/dotnet/roslyn/blob/bc0716a82fcecea45dcca285d39e6eb8fbb756c3/src/Scripting/Core/Hosting/ObjectFormatter/CommonObjectFormatter.Visitor.cs#L708-L733)